Make preserved alignment equal to PAGE_SIZE.
Make preserved alignment 4096 bytes.  Increase the number of relative
relocations in the test data so that before packing it exceeds this.
See http://www.airs.com/blog/archives/189 for motivation.

Fix a bug where sections that are not part of the process image (for
example, debug and strings data) were being inadvertently scanned for
possible relocation targets.

Small readability improvements.

BUG=385553

Review URL: https://codereview.chromium.org/434853002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@286996 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 5cd648405db836c04870f24b894eea4ba07df900

+++ tools/relocation_packer/src/delta_encoder.cc
@@ -48,7 +48,7 @@ void RelocationDeltaCodec::Decode(const std::vector<ELF::Sxword>& packed,
 
   // Ensure that the packed data offers enough pairs.  There may be zero
   // padding on it that we ignore.
-  CHECK(packed[0] <= (packed.size() - 1) >> 1);
+  CHECK(static_cast<size_t>(packed[0]) <= (packed.size() - 1) >> 1);
 
   ELF::Addr offset = 0;
   ELF::Sxword addend = 0;
